# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HealthPass Data Submission API",
    "description": "Swagger definition for HealthPass data submission API",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "/api/v1/datasubmission",
      "description": "CDT"
    }
  ],
  "paths": {
...
    "/organization/client": {
      "put": {
        "tags": [
          "organization"
        ],
        "summary": "Update organization client",
        "description": "Update organization client",
        "operationId": "updateClient",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Updated organization client",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "organization": {
                    "type": "string"
                  },
                  "clientName": {
                    "type": "string"
                  },
                  "notifyTextDataIngestResultAndroid": {
                    "type": "string"
                  },
                  "notifyTextDataIngestResultiOS": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated organization client"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/cos/{regEntity}": {
      "get": {
        "tags": [
          "cos"
        ],
        "summary": "Get file names from organization COS bucket",
        "description": "Get file names from organization COS bucket",
        "operationId": "getCOSFileNames",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "regEntity",
            "in": "path",
            "description": "Registration entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved file names from COS bucket"
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/cos/{regEntity}/owner/{holderId}": {
      "get": {
        "tags": [
          "cos"
        ],
        "summary": "Get all files from organization COS bucket for a holder",
        "description": "Get all files from organization COS bucket for a holder",
        "operationId": "getCOSFilesByHolderId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "regEntity",
            "in": "path",
            "description": "Registration entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "holderId",
            "in": "path",
            "description": "Owner",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Start date",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2021-03-20"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "End date",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2021-04-15"
            }
          },
          {
            "name": "publicKeyType",
            "in": "query",
            "description": "Type of public key",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "spki",
                "pkcs1"
              ]
            }
          },
          {
            "name": "signatureValue",
            "in": "query",
            "description": "Signature value",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all items from COS bucket for a holder"
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/cos/{regEntity}/{filename}": {
      "get": {
        "tags": [
          "cos"
        ],
        "summary": "Get file from organization COS bucket",
        "description": "Get file from organization COS bucket",
        "operationId": "getCOSFile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "regEntity",
            "in": "path",
            "description": "Registration entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "description": "File name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved file from COS bucket"
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "cos"
        ],
        "summary": "Delete file from organization COS bucket",
        "description": "Delete file from organization COS bucket",
        "operationId": "deleteCOSFile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "regEntity",
            "in": "path",
            "description": "Registration entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "description": "File name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted file from COS bucket"
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "BasicErrorMessage": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "InvalidRequestError": {
        "description": "Missing or invalid required fields in request body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "InvalidGetRequestError": {
        "description": "Invalid fields in request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Logged in user is unauthorized to make this request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
    }
  }
}